+++ Editor/PerformanceTestRunSaver.cs
@@ -1,5 +1,4 @@
-﻿#if UNITY_2018_3_OR_NEWER
-using System;
+﻿using System;
 using System.IO;
 using UnityEditor;
 using UnityEngine;
@@ -22,23 +21,23 @@ namespace Unity.PerformanceTesting.Editor
     [Serializable]
     public class PerformanceTestRunSaver : ScriptableObject, ICallbacks
     {
-        private PerformanceTestRun m_Run;
-
-#if UNITY_2019_1_OR_NEWER
         void ICallbacks.RunStarted(ITestAdaptor testsToRun)
         {
         }
 
         void ICallbacks.RunFinished(ITestResultAdaptor result)
         {
+            PlayerCallbacks.saved = false;
+
             try
             {
                 var resultWriter = new ResultsWriter();
-                string xmlPath = Path.Combine(Application.persistentDataPath, "TestResults.xml");
-                string jsonPath = Path.Combine(Application.persistentDataPath, "PerformanceTestResults.json");
+                var xmlPath = Path.Combine(Application.persistentDataPath, "TestResults.xml");
+                var jsonPath = Path.Combine(Application.persistentDataPath, "PerformanceTestResults.json");
                 resultWriter.WriteResultToFile(result, xmlPath);
                 var xmlParser = new TestResultXmlParser();
                 var run = xmlParser.GetPerformanceTestRunFromXml(xmlPath);
+                if (run == null) return;
                 File.WriteAllText(jsonPath, JsonUtility.ToJson(run, true));
             }
             catch (Exception e)
@@ -54,37 +53,5 @@ namespace Unity.PerformanceTesting.Editor
         void ICallbacks.TestFinished(ITestResultAdaptor result)
         {
         }
-#else
-        void ICallbacks.RunStarted(ITest testsToRun)
-        {
-        }
-
-        void ICallbacks.RunFinished(ITestResult result)
-        {
-            try
-            {
-                var resultWriter = new ResultsWriter();
-                string xmlPath = Path.Combine(Application.streamingAssetsPath, "TestResults.xml");
-                string jsonPath = Path.Combine(Application.streamingAssetsPath, "PerformanceTestResults.json");
-                resultWriter.WriteResultToFile(result, xmlPath);
-                var xmlParser = new TestResultXmlParser();
-                var run = xmlParser.GetPerformanceTestRunFromXml(xmlPath);
-                File.WriteAllText(jsonPath, JsonUtility.ToJson(run, true));
-            }
-            catch (Exception e)
-            {
-                Console.WriteLine(e.Message + "\n" + e.InnerException);
-            }
-        }
-
-        void ICallbacks.TestStarted(ITest test)
-        {
-        }
-
-        void ICallbacks.TestFinished(ITestResult result)
-        {
-        }
-#endif
     }
 }
-#endif