+++ Editor/TestResultXmlParser.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Xml.Linq;
+using Unity.PerformanceTesting.Runtime;
 using UnityEngine;
 
 namespace Unity.PerformanceTesting.Editor
@@ -21,12 +22,12 @@ namespace Unity.PerformanceTesting.Editor
         {
             if (string.IsNullOrEmpty(resultXmlFileName))
             {
-                throw new ArgumentNullException(resultXmlFileName, resultXmlFileName);
+                Debug.LogWarning($"Test results path is null or empty.");
             }
 
             if (!File.Exists(resultXmlFileName))
             {
-                throw new FileNotFoundException("Result file not found; {0}", resultXmlFileName);
+                Debug.LogWarning($"Test results file does not exists at path: {resultXmlFileName}");
             }
         }
 
@@ -38,18 +39,19 @@ namespace Unity.PerformanceTesting.Editor
             }
             catch (Exception e)
             {
-                var errMsg = string.Format("Failed to load xml result file: {0}", resultXmlFileName);
-                WriteExceptionConsoleErrorMessage(errMsg, e);
-                throw;
+                var errMsg = $"Failed to load xml result file: {resultXmlFileName}";
+                Debug.LogWarning($"{errMsg}\r\nException: {e.Message}\r\n{e.StackTrace}");
             }
+
+            return null;
         }
 
-        private PerformanceTestRun TryParseXmlToPerformanceTestRun(XDocument xmlDocument)
+        private PerformanceTestRun TryParseXmlToPerformanceTestRun(XContainer xmlDocument)
         {
             var output = xmlDocument.Descendants("output").ToArray();
-            if (output == null || !output.Any())
+            if (!output.Any())
             {
-                throw new Exception("The xmlDocument passed to the TryParseXmlToPerformanceTestRun method does not have any \'ouput\' xml tags needed for correct parsing.");
+                return null;
             }
 
             var run = new PerformanceTestRun();
@@ -85,18 +87,18 @@ namespace Unity.PerformanceTesting.Editor
             foreach (var element in output)
             {
                 var elements = element.Value.Split('\n');
-                if(elements.Any(e => e.Length > 0 && e.Substring(0, 2).Equals("##")))
+                if (!elements.Any(e => e.Length > 0 && e.Substring(0, 2).Equals("##"))) continue;
                 {
-                    var line = elements.First(e => e.Length > 0 && e.Substring(0, 2).Equals("##"));
+                    var lines = elements.Where(e => e.Length > 0 && e.Substring(0, 2).Equals("##"));
 
-                    var json = GetJsonFromHashtag("performancetestruninfo", line);
-
-                    // This is the happy case where we have a performancetestruninfo json object
-                    if (json != null)
+                    foreach (var line in lines)
                     {
-                        var result = TryDeserializePerformanceTestRunJsonObject(json);
-                        if (result != null)
+                        var json = GetJsonFromHashtag("performancetestruninfo", line);
+
+                        if (json != null)
                         {
+                            var result = TryDeserializePerformanceTestRunJsonObject(json);
+                            if (result == null) continue;
                             run.TestSuite = result.TestSuite;
                             run.EditorVersion = result.EditorVersion;
                             run.QualitySettings = result.QualitySettings;
@@ -105,73 +107,47 @@ namespace Unity.PerformanceTesting.Editor
                             run.PlayerSettings = result.PlayerSettings;
                             run.PlayerSystemInfo = result.PlayerSystemInfo;
                             run.StartTime = result.StartTime;
-                            // @TODO fix end time, does it matter for now?
-                            run.EndTime = run.StartTime;
+                            run.EndTime = Utils.DateToInt(DateTime.Now);
+                            break;
                         }
                     }
-                    // Unhappy case where we couldn't find a performancetestruninfo object
-                    // This could be because we have missing metadata for the test run
-                    // In this case, we try to look for a performancetestresult json object
-                    // We should have at least startime metadata  that we can use to correctly
-                    // display the test results on the x-axis of the chart
-                    else
-                    {
-                        json = GetJsonFromHashtag("performancetestresult", line);
-                        if (json != null)
-                        {
-                            var result = TryDeserializePerformanceTestRunJsonObject(json);
-                            run.StartTime = result.StartTime;
-                            // @TODO fix end time, does it matter for now?
-                            run.EndTime = run.StartTime;
-                        }
-                    }
-                } 
+                }
             }
         }
 
         private PerformanceTest TryDeserializePerformanceTestResultJsonObject(string json)
         {
-            PerformanceTest performanceTestResult;
             try
             {
-                performanceTestResult = JsonUtility.FromJson<PerformanceTest>(json);
+                return JsonUtility.FromJson<PerformanceTest>(json);
             }
             catch (Exception e)
             {
-                var errMsg = string.Format("Exception thrown while deserializing json string to PerformanceTestResult: {0}", json);
-                WriteExceptionConsoleErrorMessage(errMsg, e);
-                throw;
+                var errMsg = $"Exception thrown while deserializing json string to PerformanceTestResult: {json}";
+                Debug.LogWarning($"{errMsg}\r\nException: {e.Message}\r\n{e.StackTrace}");
             }
 
-            return performanceTestResult;
+            return null;
         }
-
-        private void WriteExceptionConsoleErrorMessage(string errMsg, Exception e)
-        {
-            Console.Error.WriteLine("{0}\r\nException: {1}\r\nInnerException: {2}", errMsg, e.Message,
-                e.InnerException.Message);
-        }
-
+        
         private PerformanceTestRun TryDeserializePerformanceTestRunJsonObject(string json)
         {
-            PerformanceTestRun performanceTestRun;
             try
             {
-                performanceTestRun = JsonUtility.FromJson<PerformanceTestRun>(json);
+                return JsonUtility.FromJson<PerformanceTestRun>(json);
             }
             catch (Exception e)
             {
-                var errMsg = string.Format("Exception thrown while deserializing json string to PerformanceTestRun: {0}", json);
-                WriteExceptionConsoleErrorMessage(errMsg, e);
-                throw;
+                var errMsg = $"Exception thrown while deserializing json string to PerformanceTestRun: {json}";
+                Debug.LogWarning($"{errMsg}\r\nException: {e.Message}\r\n{e.StackTrace}");
             }
 
-            return performanceTestRun;
+            return null;
         }
 
         private string GetJsonFromHashtag(string tag, string line)
         {
-            if (!line.Contains(string.Format("##{0}:",tag))) return null;
+            if (!line.Contains($"##{tag}:")) return null;
             var jsonStart = line.IndexOf('{');
             var openBrackets = 0;
             var stringIndex = jsonStart;