+++ Runtime/Attributes/PerformanceAttribute.cs
@@ -1,12 +1,18 @@
 ﻿using System;
 using System.Collections;
+using NUnit.Framework;
 using NUnit.Framework.Interfaces;
 using UnityEngine.TestTools;
 
 namespace Unity.PerformanceTesting
 {
-    public class PerformanceAttribute : Attribute, IOuterUnityTestAction
+    public class PerformanceAttribute : CategoryAttribute, IOuterUnityTestAction
     {
+        public PerformanceAttribute() : base("Performance")
+        {
+        }
+
+
         public IEnumerator BeforeTest(ITest test)
         {
             PerformanceTest.StartTest(test);