+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [1.3.3] - 2019-12-12
+
+Fix metadata reporting for editor tests
+Fix configuration fields in json report
+Method measurements will now measure the time for a single iteration
+Fix warning after cleaning resources
+Fix test suite when running in the editor
+Remove metadata collectors tests
+Switch to errors from exceptions when parsing results
+Increase minimum unity version to 2019.3
+
+### Categorize performance tests as `performance`
+
+Categorize performance tests as performance
+Remove profiler section on docs as the feature was removed
+ProfilerMarkers can now be called with string params
+Switch measuring frames and methods to stopwatch
 ## [1.0.4-preview] - 2020-02-18
 
 ### This is the first release of *com.unity.xr.test-framework.performance*.